+++ src/backend/template_selector.py
@@ -2,13 +2,46 @@ from os import PathLike
 from pathlib import Path
 
 from src.backend.utils.working_dir import RUNTIME_DIR
+from src.enums.media_type import MediaType
 
-DEFAULT_TEMPLATE = """\
+DEF_MV_TEMPLATE = """\
 Info
-Title:                  : {{ movie_title }} {{ release_year_parentheses }}
+Title:                  : {{ title_exact }} {{ release_year_parentheses }}
+Source:                 : {{ source_file_no_ext }}
 Format Profile          : {{ format_profile }}
 Resolution              : {{ resolution }}
-Average Bitrate         : {{ mi_video_bit_rate }}
+Average Bitrate         : {{ video_bit_rate }}
+{% if releasers_name %}
+Encoder                 : {{ releasers_name }}
+{% endif %}
+
+{% if screen_shots %}
+{{ screen_shots }}
+{% endif %}
+
+{% if release_notes %}
+Release Notes:
+{{ release_notes }}
+{% endif %}
+
+{% if media_info_short %}
+MediaInfo
+{{ media_info_short }}
+{% endif %}
+
+{{ shared_with_bbcode }}"""
+
+DEF_SERIES_TEMPLATE = """\
+Info
+Title:                  : {{ title_exact }} {{ release_year_parentheses }}
+Season                  : {{ season_number }}
+Episode                 : {{ episode_number }}
+Episode Title           : {{ episode_title_exact }}
+Air Date                : {{ air_date }}
+Source                  : {{ source_file_no_ext }}
+Format Profile          : {{ format_profile }}
+Resolution              : {{ resolution }}
+Average Bitrate         : {{ video_bit_rate }}
 {% if releasers_name %}
 Encoder                 : {{ releasers_name }}
 {% endif %}
@@ -74,9 +107,14 @@ class TemplateSelectorBackEnd:
             with open(_path, "r", encoding="utf-8") as template:
                 return template.read()
 
-    def create_template(self, path: PathLike[str] | str) -> Path:
+    def create_template(self, path: PathLike[str] | str, media_type: MediaType) -> Path:
         with open(path, "w", encoding="utf-8") as new_template:
-            new_template.write(self.get_default_template())
+            # determine correct template based on media type selection from frontend
+            new_template.write(
+                DEF_MV_TEMPLATE
+                if media_type is MediaType.MOVIE
+                else DEF_SERIES_TEMPLATE
+            )
         self.load_templates()
         return Path(path)
 
@@ -87,6 +125,3 @@ class TemplateSelectorBackEnd:
     def delete_template(self, path: PathLike[str] | str) -> None:
         Path(path).unlink()
         self.load_templates()
-
-    def get_default_template(self) -> str:
-        return DEFAULT_TEMPLATE